requirements.txt
moviepy
openai-whisper
